move timer into demo
